+++ provisioning/scripts/common/common/metadata.sh
@@ -3,9 +3,8 @@
 # Custom (All)
 
 # This file relies on environment variables to provide template information. Required variables are:
-# * TEMPLATE_NAME      equal to the 'proper' name of the template (e.g. antarctica/centos7)
-# * TEMPLATE_NAME_ALT  equal to the 'safe' name of the template   (e.g. antarctica-centos7)
-# * TEMPLATE_VERSION   equal to the version of the template       (e.g. 1.2.3)
+# * TEMPLATE_NAME      equal to the name of the template (e.g. antarctica-centos7)
+# * TEMPLATE_VERSION   equal to the version of the template (e.g. 2012-07-20)
 
 # Create local facts directory
 mkdir -p /etc/ansible/facts.d;
@@ -13,8 +12,7 @@ mkdir -p /etc/ansible/facts.d;
 # Create fact file with template information - Produces a file like:
 # ```
 # [general]
-# name=antarctica/centos7
-# name_alt=antarctica-centos7
-# version=1.2.3
+# name=antarctica-centos7
+# version=2012-07-20
 # ````
-printf "[general]\nname=$TEMPLATE_NAME\nname_alt=$TEMPLATE_NAME_ALT\nversion=$TEMPLATE_VERSION\n" > /etc/ansible/facts.d/os_template.fact;
+printf "[general]\nname=$TEMPLATE_NAME\nversion=$TEMPLATE_VERSION\n" > /etc/ansible/facts.d/os_template.fact;