+++ provisioning/scripts/centos/common/sudoers.sh
@@ -1,5 +1,0 @@
-#!/usr/bin/env bash -eux
-
-# Custom (section)
-
-sed -i -e 's/# %wheel\tALL=(ALL)\tNOPASSWD: ALL/%wheel\tALL=(ALL)\tNOPASSWD: ALL/' /etc/sudoers