+++ provisioning/scripts/common/common/sshd.sh
@@ -1,6 +1,0 @@
-#!/usr/bin/env bash -eux
-
-# Bento (All)
-
-echo "UseDNS no" >> /etc/ssh/sshd_config
-echo "GSSAPIAuthentication no" >> /etc/ssh/sshd_config