+++ provisioning/terraform/10-null_resource.tf
@@ -1,0 +1,4 @@
+#
+# This file is used to ensure a state file is created so that Terraform outputs can be referenced in other projects
+
+resource "null_resource" "null" {}