+++ packer/provisioning/scripts/centos/desktop/cleanup.sh
@@ -1,0 +1,52 @@
+#!/usr/bin/env bash -eux
+
+# Bento (All)
+
+# Clean up network interface persistence
+rm -f /etc/udev/rules.d/70-persistent-net.rules;
+mkdir -p /etc/udev/rules.d/70-persistent-net.rules;
+rm -f /lib/udev/rules.d/75-persistent-net-generator.rules;
+rm -rf /dev/.udev/;
+
+for ndev in `ls -1 /etc/sysconfig/network-scripts/ifcfg-*`; do
+    if [ "`basename $ndev`" != "ifcfg-lo" ]; then
+        sed -i '/^HWADDR/d' "$ndev";
+        sed -i '/^UUID/d' "$ndev";
+    fi
+done
+
+for ndev in `ls -1 /etc/sysconfig/network-scripts/ifcfg-*`; do
+    if [ "`basename $ndev`" != "ifcfg-lo" ]; then
+        sed -i '/^HWADDR/d' "$ndev";
+        sed -i '/^UUID/d' "$ndev";
+    fi
+done
+
+# new-style network device naming for centos7 - radio off & remove all interface configration
+nmcli radio all off
+/bin/systemctl stop NetworkManager.service
+for ifcfg in `ls /etc/sysconfig/network-scripts/ifcfg-* |grep -v ifcfg-lo` ; do
+rm -f $ifcfg
+done
+rm -rf /var/lib/NetworkManager/*
+
+echo "==> Setup /etc/rc.d/rc.local for EL7"
+cat <<_EOF_ | cat >> /etc/rc.d/rc.local
+#BENTO-BEGIN
+LANG=C
+# delete all connection
+for con in \`nmcli -t -f uuid con\`; do
+  if [ "\$con" != "" ]; then
+    nmcli con del \$con
+  fi
+done
+# add gateway interface connection.
+gwdev=\`nmcli dev | grep ethernet | egrep -v 'unmanaged' | head -n 1 | awk '{print \$1}'\`
+if [ "\$gwdev" != "" ]; then
+  nmcli connection add type ethernet ifname \$gwdev con-name \$gwdev
+fi
+sed -i "/^#BENTO-BEGIN/,/^#BENTO-END/d" /etc/rc.d/rc.local
+chmod -x /etc/rc.d/rc.local
+#BENTO-END
+_EOF_
+chmod +x /etc/rc.d/rc.local