+++ usage/examples/terraform/bas-development-cluster/single-instance/00-terraform.tf
@@ -1,0 +1,8 @@
+#
+# This file is used to define Terraform core resources
+
+# These resources are defined as complete examples designed for use in other projects
+
+terraform {
+  required_version = "~> 0.12"
+}