+++ CHANGELOG.md
@@ -5,6 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased][unreleased]
 
+### Fixed
+
+* Update template and provisioning scripts to remove 'artefact' and 'artefact_alt' identifiers in favour of a single 'template' identifier
+
 ## 0.1.0 (2018-09-01)
 
 ### Added