+++ CHANGELOG.md
@@ -8,6 +8,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Fixed
 
 * Update template and provisioning scripts to remove 'artefact' and 'artefact_alt' identifiers in favour of a single 'template' identifier
+### Changed
+
+* Pinning Packer Docker image version to prevent issues with regressions added in the latest releases
 
 ## 0.1.0 (2018-09-01)