+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased][unreleased]
 
+## 0.2.0 (2020-05-18)
+
 ### Removed [BRAKING!]
 
 * Vagrant insecure key removed from desktop providers of the `antarctica-centos7` image
@@ -28,7 +30,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ### Fixed
 
 * Correctly disabling SE Linux in the `antarctica-centos7` Digital Ocean image
-* Update template and provisioning scripts to remove 'artefact' and 'artefact_alt' identifiers in favour of a single 
+* Update template and provisioning scripts to remove 'artefact' and 'artefact_alt' identifiers in favour of a single
   'template' identifier
 * Pinning Packer version in CI